mod: rename ValidateRequireValue class
